lsp-devtools: move LiveSqlHandler to utils module
